scraper_service/requirements.txt
fastapi
uvicorn[standard]
scrapegraphai
python-dotenv
